Reorganize screens directory
