+++ adc_app/lib/screens/client/client_signup_screen.dart
@@ -1,34 +1,37 @@
-import 'package:adc_app/screens/applications/doula_app.dart';
 import 'package:flutter/material.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:adc_app/theme/colors.dart';
 import 'package:flutter/services.dart';
-import 'package:adc_app/util/auth.dart';
-import 'package:adc_app/models/doula.dart';
+import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:adc_app/screens/home_screen.dart';
+import 'package:adc_app/theme/colors.dart';
+import 'package:adc_app/util/auth.dart';
+import 'package:adc_app/screens/applications/test_app.dart';
+import 'package:adc_app/models/client.dart';
+import 'package:adc_app/screens/applications/client/client_app.dart';
 
 
-class DoulaSignupPage extends StatefulWidget {
-  DoulaSignupPage({Key key}) : super(key: key);
+class ClientSignupPage extends StatefulWidget {
+  ClientSignupPage({Key key}) : super(key: key);
 
   @override
-  State<StatefulWidget> createState() => _DoulaSignupPageState();
+  State<StatefulWidget> createState() => _ClientSignupPageState();
 }
 
-class _DoulaSignupPageState extends State<DoulaSignupPage> {
+class _ClientSignupPageState extends State<ClientSignupPage> {
   final GlobalKey<FormState> _registerFormKey = GlobalKey<FormState>();
   bool _autoValidate = false;
   TextEditingController _emailInputController;
   TextEditingController _pwdInputController;
   TextEditingController _confirmPwdInputController;
+  String _emailTakenError;
 
   bool _passwordVisible;
-  String _emailTakenError;
 
   String userId;
   Key key;
 
   final MenuMaker _myMenuMaker = MenuMaker();
+
 
   @override
   void initState() {
@@ -47,11 +50,12 @@ class _DoulaSignupPageState extends State<DoulaSignupPage> {
     _registerFormKey.currentState.reset();
   }
 
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
         appBar: AppBar(
-          title: Text("Apply to be an ADC Doula"),
+          title: Text("Request a Doula"),
         ),
         drawer: _myMenuMaker.createMenu(context),
         body: Container(
@@ -79,12 +83,18 @@ class _DoulaSignupPageState extends State<DoulaSignupPage> {
                   textColor: Colors.white,
                   padding: EdgeInsets.all(15.0),
                   child: Text("LOG IN")),
+              SizedBox(
+                height: 5,
+              ),
             ],
           )),
         ));
   }
 
   Widget _registerForm() {
+
+
+
     return Form(
         key: _registerFormKey,
         autovalidate: _autoValidate,
@@ -108,12 +118,12 @@ class _DoulaSignupPageState extends State<DoulaSignupPage> {
                 suffixIcon: IconButton(
                   icon: Icon(
                     // Based on passwordVisible state choose the icon
-                      _passwordVisible
-                          ? Icons.visibility
-                          : Icons.visibility_off,
-                      color: _passwordVisible
-                          ? themeColors["black"]
-                          : themeColors["coolGray5"]
+                    _passwordVisible
+                        ? Icons.visibility
+                        : Icons.visibility_off,
+                    color: _passwordVisible
+                        ? themeColors["black"]
+                        : themeColors["coolGray5"]
                   ),
                   onPressed: () {
                     setState(() {
@@ -156,6 +166,9 @@ class _DoulaSignupPageState extends State<DoulaSignupPage> {
                 if (form.validate()) {
                   form.save();
                   try {
+//                    userId = await widget.auth.signUp(
+//                        _emailInputController.text, _pwdInputController.text);
+
                     AuthResult result = await FirebaseAuth.instance
                         .createUserWithEmailAndPassword(
                             email: _emailInputController.text.toString().trim(),
@@ -166,24 +179,24 @@ class _DoulaSignupPageState extends State<DoulaSignupPage> {
                     print("sign up returned user id: $userId");
 
                     if (userId.length > 0 && userId != null) {
-                      Doula doulaApplicant = new Doula(userId, "doula",
+                      Client clientApplicant = new Client(userId, "client",
                           _emailInputController.text.toString().trim());
+                      print("new client: ${clientApplicant.toString()}");
                       Navigator.push(
                         context,
                         MaterialPageRoute(
-                            builder: (context) =>
-                                DoulaAppPage(key: key, user: doulaApplicant)),
+                            builder: (context) => ClientAppPersonalInfoPage(
+                                key: key, user: clientApplicant)),
                       );
-                    } else {
-                      print("invalid user id returned from firebase");
                     }
-                  } catch (e) {
+                  }
+                  catch (e) {
                     if (e is PlatformException && e.code == "ERROR_EMAIL_ALREADY_IN_USE") {
                       setState(() {
                         _emailTakenError = e.message;
                       });
                     }
-                    print("Doula account sign up error: $e");
+                    print("Client account sign up error: $e");
                     form.reset();
                   }
                 }
