+++ adc_app/lib/screens/messaging/text_bank.dart
